added contract tests for all bhce v2 auth resources
